# app/api/routes.py
from __future__ import annotations

import mimetypes
from email.utils import formatdate, parsedate_to_datetime
from http import HTTPStatus

from fastapi import APIRouter, HTTPException, Request, Response
from fastapi.responses import FileResponse

from app.models.schemas import (
    DeleteImageResponse,
    ImageMetadataResponse,
    ImagesResponse,
    RenameDirectoryRequest,
    RenameDirectoryResponse,
    SubdirectoriesResponse,
    TagIndexRefreshJobStartResponse,
    TagIndexRefreshJobStatusResponse,
    TagIndexRefreshResponse,
    TagListResponse,
    TagQueryRequest,
    TagQueryResponse,
    TagSummaryEntry,
)
from app.models.types import DirectoryId, FileId
from app.services.image_service import (
    ConflictError,
    ImageService,
    ResourceNotFoundError,
    ServiceError,
    UnsupportedMediaTypeError,
    ValidationError,
)
from app.services.tag_index_service import TagIndexService


def create_api_router(service: ImageService, tag_index_service: TagIndexService) -> APIRouter:
    router = APIRouter(prefix="/api")

    @router.get(
        "/subdirectories",
        response_model=SubdirectoriesResponse,
        operation_id="listSubdirectories",
    )
    def get_subdirectories() -> SubdirectoriesResponse:
        return SubdirectoriesResponse(subdirectories=service.list_subdirectories())

    @router.get(
        "/images/{directory_id}",
        response_model=ImagesResponse,
        operation_id="listImages",
    )
    def get_images(directory_id: DirectoryId) -> ImagesResponse:
        try:
            directory, images = service.list_images(directory_id)
        except ResourceNotFoundError as exc:
            raise HTTPException(status_code=HTTPStatus.NOT_FOUND) from exc
        return ImagesResponse(directory_id=directory_id, subdirectory=directory.name, images=images)

    def _build_image_response(file_id: FileId, request: Request, include_body: bool) -> Response:
        try:
            file_path = service.resolve_image(file_id)
        except ResourceNotFoundError as exc:
            raise HTTPException(status_code=HTTPStatus.NOT_FOUND) from exc
        except UnsupportedMediaTypeError as exc:
            raise HTTPException(status_code=HTTPStatus.UNSUPPORTED_MEDIA_TYPE) from exc

        try:
            stat_result = file_path.stat()
        except OSError as exc:
            raise HTTPException(status_code=HTTPStatus.INTERNAL_SERVER_ERROR) from exc

        etag = f'W/"{stat_result.st_mtime_ns}-{stat_result.st_size}"'
        last_modified = formatdate(stat_result.st_mtime, usegmt=True)
        cache_headers = {
            "ETag": etag,
            "Last-Modified": last_modified,
            "Cache-Control": "public, max-age=31536000",
        }

        if_none_match = request.headers.get("if-none-match")
        if if_none_match and if_none_match.strip() == etag:
            return Response(status_code=HTTPStatus.NOT_MODIFIED, headers=cache_headers)

        if_modified_since = request.headers.get("if-modified-since")
        if if_modified_since:
            try:
                since_timestamp = parsedate_to_datetime(if_modified_since).timestamp()
                if stat_result.st_mtime <= since_timestamp:
                    return Response(status_code=HTTPStatus.NOT_MODIFIED, headers=cache_headers)
            except (TypeError, ValueError, OverflowError):
                pass

        content_type = mimetypes.guess_type(file_path.name)[0] or "application/octet-stream"
        headers = {**cache_headers, "Content-Length": str(stat_result.st_size)}

        if not include_body:
            return Response(status_code=HTTPStatus.OK, headers=headers, media_type=content_type)

        return FileResponse(path=file_path, media_type=content_type, headers=headers)

    @router.get("/image/{file_id}", operation_id="getImage")
    def get_image(file_id: FileId, request: Request) -> Response:
        return _build_image_response(file_id=file_id, request=request, include_body=True)

    @router.head("/image/{file_id}", operation_id="headImage")
    def head_image(file_id: FileId, request: Request) -> Response:
        return _build_image_response(file_id=file_id, request=request, include_body=False)

    @router.get(
        "/image-meta/{file_id}",
        response_model=ImageMetadataResponse,
        operation_id="getImageMetadata",
    )
    def get_image_metadata(file_id: FileId) -> ImageMetadataResponse:
        try:
            directory_id, directory_name, image = service.get_image_metadata(file_id)
        except ResourceNotFoundError as exc:
            raise HTTPException(status_code=HTTPStatus.NOT_FOUND) from exc
        except UnsupportedMediaTypeError as exc:
            raise HTTPException(status_code=HTTPStatus.UNSUPPORTED_MEDIA_TYPE) from exc

        return ImageMetadataResponse(
            file_id=image.file_id,
            name=image.name,
            directory_id=directory_id,
            directory_name=directory_name,
        )

    @router.delete(
        "/image/{file_id}",
        response_model=DeleteImageResponse,
        operation_id="deleteImage",
    )
    def delete_image(file_id: FileId) -> DeleteImageResponse:
        try:
            deleted_file = service.delete_image(file_id)
        except ResourceNotFoundError as exc:
            raise HTTPException(status_code=HTTPStatus.NOT_FOUND) from exc
        except UnsupportedMediaTypeError as exc:
            raise HTTPException(status_code=HTTPStatus.UNSUPPORTED_MEDIA_TYPE) from exc
        except ServiceError as exc:
            raise HTTPException(status_code=HTTPStatus.INTERNAL_SERVER_ERROR) from exc

        return DeleteImageResponse(deleted=deleted_file.name, file_id=file_id)

    @router.put(
        "/subdirectories/{directory_id}",
        response_model=RenameDirectoryResponse,
        operation_id="renameSubdirectory",
    )
    def rename_subdirectory(directory_id: DirectoryId, payload: RenameDirectoryRequest) -> RenameDirectoryResponse:
        try:
            new_directory_id, renamed_from, renamed_to = service.rename_subdirectory(directory_id, payload.new_name)
        except ResourceNotFoundError as exc:
            raise HTTPException(status_code=HTTPStatus.NOT_FOUND) from exc
        except ValidationError as exc:
            raise HTTPException(status_code=HTTPStatus.BAD_REQUEST) from exc
        except ConflictError as exc:
            raise HTTPException(status_code=HTTPStatus.CONFLICT) from exc
        except ServiceError as exc:
            raise HTTPException(status_code=HTTPStatus.INTERNAL_SERVER_ERROR) from exc

        return RenameDirectoryResponse(
            directory_id=new_directory_id,
            renamed_from=renamed_from,
            renamed_to=renamed_to,
        )


    @router.post(
        "/tag-index/refresh-jobs",
        response_model=TagIndexRefreshJobStartResponse,
        operation_id="startTagIndexRefreshJob",
    )
    def start_tag_index_refresh_job() -> TagIndexRefreshJobStartResponse:
        job_id = tag_index_service.start_refresh_job()
        return TagIndexRefreshJobStartResponse(job_id=job_id)

    @router.get(
        "/tag-index/refresh-jobs/{job_id}",
        response_model=TagIndexRefreshJobStatusResponse,
        operation_id="getTagIndexRefreshJob",
    )
    def get_tag_index_refresh_job(job_id: str) -> TagIndexRefreshJobStatusResponse:
        status = tag_index_service.get_refresh_job_status(job_id)
        if status is None:
            raise HTTPException(status_code=HTTPStatus.NOT_FOUND)

        return TagIndexRefreshJobStatusResponse(
            job_id=status.job_id,
            status=status.status,
            progress_phase=status.progress_phase,
            processed_files=status.processed_files,
            total_files=status.total_files,
            indexed_files=status.indexed_files,
            indexed_tags=status.indexed_tags,
            counters={
                "added": status.counters.added,
                "modified": status.counters.modified,
                "deleted": status.counters.deleted,
                "reindexed": status.counters.reindexed,
            },
            error=status.error,
        )

    @router.post(
        "/tag-index/refresh",
        response_model=TagIndexRefreshResponse,
        operation_id="refreshTagIndex",
    )
    def refresh_tag_index() -> TagIndexRefreshResponse:
        """Backward-compatible synchronous refresh API.

        New clients should migrate to POST /api/tag-index/refresh-jobs and
        poll GET /api/tag-index/refresh-jobs/{job_id}. This endpoint currently
        waits for job completion and returns the final counts.
        """
        job_id = tag_index_service.start_refresh_job()
        status = tag_index_service.wait_for_job(job_id)
        if status.status == "failed":
            raise HTTPException(status_code=HTTPStatus.INTERNAL_SERVER_ERROR, detail=status.error)

        return TagIndexRefreshResponse(
            indexed_files=status.indexed_files,
            indexed_tags=status.indexed_tags,
        )


    @router.get(
        "/tag-index/tags",
        response_model=TagListResponse,
        operation_id="listTagIndexTags",
    )
    def list_tag_index_tags() -> TagListResponse:
        tags = [TagSummaryEntry(tag=tag, count=count) for tag, count in tag_index_service.list_tags()]
        return TagListResponse(tags=tags)

    @router.post(
        "/tag-index/query",
        response_model=TagQueryResponse,
        operation_id="queryTagIndex",
    )
    def query_tag_index(payload: TagQueryRequest) -> TagQueryResponse:
        file_ids = tag_index_service.query(payload.tags, payload.mode)
        return TagQueryResponse(file_ids=file_ids, total=len(file_ids))

    return router